requirements.txt
pygame==2.0.0.dev10
PyYAML==5.4.1
